server/requirements.txt
Django>=5.2.0
djangorestframework>=3.16.0
psycopg2-binary>=2.9.6
djangorestframework-simplejwt>=5.3.0
redis>=4.5.4
celery>=5.3.0
pika>=1.3.1
django-cors-headers>=4.0.0
python-dotenv>=1.0.0
gunicorn>=20.1.0 
Pillow>=10.3.0
stripe>=12.1.0
geopy>=2.4.1
django-allauth>=0.54.0
msal==1.28.0
requests==2.32.3
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0